refactor + tests
