+++ RepairsApi/V2/Controllers/Parameters/WorkOrderSearchParameters.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace RepairsApi.V2.Controllers
+{
+    public class WorkOrderSearchParameters
+    {
+        public static int MaxPageSize { get; } = 50;
+        public string PropertyReference { get; set; }
+        public string ContractorReference { get; set; }
+        public int PageNumber { get; set; } = 1;
+        private int _pageSize = 10;
+        public int PageSize
+        {
+            get => _pageSize;
+            set => _pageSize = Math.Clamp(value, 0, MaxPageSize);
+        }
+    }
+}
