+++ RepairsApi/V2/Controllers/RepairsController.cs
@@ -134,19 +134,4 @@ namespace RepairsApi.V2.Controllers
 
     }
 
-    public class WorkOrderSearchParameters
-    {
-        private const int _maxPageSize = 50;
-        public string PropertyReference { get; set; }
-        public string ContractorReference { get; set; }
-
-        public int PageNumber { get; set; } = 1;
-
-        private int _pageSize = 10;
-        public int PageSize
-        {
-            get => _pageSize;
-            set => _pageSize = value > _maxPageSize ? _maxPageSize : value;
-        }
-    }
 }
